+++ NutriCare/app/src/main/res/layout/activity_main.xml
@@ -4,31 +4,40 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".MainActivity"
+    android:background="@drawable/fondo">
 
-    <ImageView
-        android:id="@+id/ivBackground"
-        android:layout_width="wrap_content"
-        android:layout_height="0dp"
-        android:background="@drawable/fondo"
+    <Button
+        android:id="@+id/btCliente"
+        android:layout_width="250dp"
+        android:layout_height="50dp"
+        android:layout_marginTop="52dp"
+        android:layout_marginEnd="72dp"
+        android:backgroundTint="#7fa208"
+        android:text="@string/entrar_como_cliente"
+        app:cornerRadius="25dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.0" />
+        app:layout_constraintVertical_bias="0.386" />
 
-    <ImageView
-        android:id="@+id/ivLogo"
-        android:layout_width="119dp"
-        android:layout_height="117dp"
-        android:layout_marginTop="70dp"
-        android:layout_marginEnd="140dp"
+    <Button
+        android:id="@+id/btNutricionista"
+        android:layout_width="250dp"
+        android:layout_height="50dp"
+        android:layout_marginTop="52dp"
+        android:layout_marginEnd="68dp"
+        android:backgroundTint="#7fa208"
+        android:text="@string/entrar_como_nutricionista"
+        android:textSize="13sp"
+        app:cornerRadius="25dp"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/logo" />
+        app:layout_constraintVertical_bias="0.519" />
 
     <TextView
-        android:id="@+id/tvAppName"
+        android:id="@+id/tvAppName2"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="212dp"
@@ -39,83 +48,17 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <Button
-        android:id="@+id/btAsesorias"
-        android:layout_width="250dp"
-        android:layout_height="50dp"
-        android:layout_marginTop="52dp"
-        android:layout_marginEnd="116dp"
-        android:backgroundTint="#7fa208"
-        android:text="@string/asesorias"
-        app:cornerRadius="25dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/tvAppName" />
-
-    <Button
-        android:id="@+id/btHistoNutric"
-        android:layout_width="250dp"
-        android:layout_height="50dp"
-        android:layout_marginEnd="84dp"
-        android:backgroundTint="#7fa208"
-        android:text="@string/historial_nutricional"
-        app:cornerRadius="25dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btAsesorias" />
-
-    <Button
-        android:id="@+id/btDietRece"
-        android:layout_width="250dp"
-        android:layout_height="50dp"
-        android:layout_marginEnd="40dp"
-        android:backgroundTint="#7fa208"
-        android:text="@string/dietas_recetas"
-        app:cornerRadius="25dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btHistoNutric" />
-
-    <Button
-        android:id="@+id/btPerfil"
-        android:layout_width="250dp"
-        android:layout_height="50dp"
-        android:layout_marginEnd="80dp"
-        android:backgroundTint="#7fa208"
-        android:text="@string/perfil"
-        app:cornerRadius="25dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btDietRece" />
-
-    <Button
-        android:id="@+id/btInform"
-        android:layout_width="250dp"
-        android:layout_height="50dp"
-        android:layout_marginEnd="116dp"
-        android:backgroundTint="#7fa208"
-        android:text="@string/informate_mas"
-        app:cornerRadius="25dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btPerfil" />
-
-    <Button
-        android:id="@+id/btCitas"
-        android:layout_width="90dp"
-        android:layout_height="40dp"
-        android:layout_marginEnd="16dp"
-        android:layout_marginBottom="30dp"
-        android:backgroundTint="#7fa208"
-        android:text="@string/citas"
-        android:textSize="14sp"
-        app:cornerRadius="25dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
     <ImageView
-        android:id="@+id/imageView4"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:layout_marginTop="24dp"
-        android:layout_marginEnd="24dp"
+        android:id="@+id/imageView6"
+        android:layout_width="127dp"
+        android:layout_height="118dp"
+
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/enter" />
+        app:layout_constraintVertical_bias="0.112"
+
+        app:srcCompat="@drawable/logo" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
